% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea2.r
\name{call_to_list}
\alias{call_to_list}
\title{call_to_list}
\usage{
call_to_list(call_url)
}
\arguments{
\item{call_url}{the API url that will be sent to the bea.gov API}
}
\description{
These functions need to have the httr package to make the GET request
against the bea.gov API
}
