% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data.R
\docType{data}
\name{census_data}
\alias{census_data}
\title{A data set of gdp per state per year}
\format{
Data frame with columns
\describe{
\item{Year}{Year of data observation}
\item{gdp_per_state_per_year}{As the name suggests, GDP per state per year.}
\item{state}{Full state name.}

}
}
\usage{
census_data
}
\description{
A data set of gdp per state per year
}
\examples{
  census_data
}
\keyword{datasets}
