% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime_functions.R
\name{clean_data}
\alias{clean_data}
\title{clean_data}
\usage{
clean_data(data, year)
}
\arguments{
\item{data}{calls the data from the get_url function and cleans that data frame so its ready for the user}

\item{year}{calls the year from the get_url function and used it in \code{if} statements to execute certain code based on the year the data is on}
}
\description{
clean_data
}
