% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime_data.R
\docType{data}
\name{crime_data}
\alias{crime_data}
\title{Different crime counts per state per year}
\format{
Data frame with columns
\describe{
\item{state}{Full state name. }
\item{Year}{Year of data observation}
\item{Overall Homeless}{The total nubmer of homeless people in that state for that year.}
}
}
\usage{
crime_data
}
\description{
Different crime counts per state per year
}
\examples{
  crime_data
}
\keyword{datasets}
