% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cen.R
\name{establish_census_api}
\alias{establish_census_api}
\title{enter api key}
\usage{
establish_census_api(apikey)
}
\arguments{
\item{apikey, }{your personal api key, you must ask for this on the census website}
}
\description{
enter api key
}
