% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea2.r
\name{filter_year}
\alias{filter_year}
\title{filter_year}
\usage{
filter_year(data, start = 0, end = 9999)
}
\arguments{
\item{data}{\itemize{
\item The data from the bea.gov API that needs to be filtered to the specified years
}}

\item{start}{\itemize{
\item The starting year for the dataset. Default is set to 0000.
}}

\item{end}{\itemize{
\item Then ending year for the dataset. Default is set to 9999.
}}
}
\value{
Dataset filtered to the specified years.
}
\description{
filter_year
}
