% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_data.R
\docType{data}
\name{full_data}
\alias{full_data}
\title{A collection of all of our homelessness adjacent data sets.}
\format{
Data frame with columns
\describe{
\item{state}{Full state name. }
\item{Year}{Year of data observation}
\item{Overall Homeless}{The total nubmer of homeless people in that state for that year.}
}
}
\usage{
full_data
}
\description{
A collection of all of our homelessness adjacent data sets.
}
\examples{
  full_data
}
\keyword{datasets}
