% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cen.R
\name{get_census_data}
\alias{get_census_data}
\title{download census data}
\usage{
get_census_data(
  table = "B01001",
  start_year = 2010,
  end_year = 2019,
  geography = "state",
  survey = "acs1"
)
}
\arguments{
\item{table, }{the name of the table}

\item{start_year}{the first year you would like data from}

\item{end_year}{the latest year you would like data from}

\item{geography}{state, county, ect.}

\item{survey}{acs1 or acs5}
}
\value{
the data set
}
\description{
download census data
}
