% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud.r
\name{get_hud_data}
\alias{get_hud_data}
\title{download hud data}
\usage{
get_hud_data(
  y,
 
    url1 = "https://www.huduser.gov/portal/sites/default/files/xls/2007-2021-PIT-Counts-by-State.xlsx"
)
}
\arguments{
\item{y}{\itemize{
\item The year that you want to find data. Current avalable years are between 2007 - 2020
}}

\item{url1}{\itemize{
\item The URL pulls from hud dataset of number of homelessness by state
}}
}
\description{
download hud data
}
