% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime_functions.R
\name{get_url}
\alias{get_url}
\title{get_url}
\usage{
get_url(start_year = 2006, end_year = 2017)
}
\arguments{
\item{year}{year of data desired from 2006-2019 excluding 2008 and 2016}
}
\description{
get_url
}
\author{
Becca Ebersole
}
