% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud.r, R/hud_data.r
\docType{data}
\name{hud_data}
\alias{hud_data}
\title{download hud data}
\format{
Data frame with columns
\describe{
\item{state}{Full state name. }
\item{Year}{Year of data observation}
\item{Overall Homeless}{The total nubmer of homeless people in that state for that year.}
}
}
\usage{
hud_data(
  y,
 
    url1 = "https://www.huduser.gov/portal/sites/default/files/xls/2007-2021-PIT-Counts-by-State.xlsx"
)

hud_data
}
\arguments{
\item{y}{\itemize{
\item The year that you want to find data. Current avalable years are between 2007 - 2020
}}

\item{url1}{\itemize{
\item The URL pulls from hud dataset of number of homelessness by state
}}
}
\description{
download hud data

Number of Homeless people per state per year.
}
\examples{
  hud_data
}
\keyword{datasets}
