% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea.r
\name{make_get_url}
\alias{make_get_url}
\title{make_get_url}
\usage{
make_get_url(dataset_name = "", dataset = "", line_code = "", api_key = "")
}
\arguments{
\item{dataset_name}{\itemize{
\item The macro type of dataset. For most of the bea functions, the default is Regional
}}

\item{dataset}{\itemize{
\item the section of the data for the bea data
}}

\item{line_code}{\itemize{
\item this is the specific slice of data that you would like
}}

\item{api_key}{\itemize{
\item Required for all data calls to the bea website.
}}
}
\value{
An API url that can be used to make a call against the bea.gov API.
}
\description{
make_get_url
}
