% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea2.r
\name{tot_employ_bea}
\alias{tot_employ_bea}
\title{tot_employ_bea}
\usage{
tot_employ_bea(api_key = "", start_year = 0, end_year = 9999)
}
\arguments{
\item{api_key}{\itemize{
\item The API key from the bea.gov website. The following link will take you to the bea.gov website to register for an API key: https://apps.bea.gov/API/signup/index.cfm
}}

\item{start_year}{The year that the data should start. The earliest data is from 2000.}

\item{end_year}{The year that the data should end. The latest data is from 2020.}
}
\value{
The number of people that were employed for a year in a state to the desired years
}
\description{
tot_employ_bea
}
