% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_employment_data.R
\docType{data}
\name{total_employment_data}
\alias{total_employment_data}
\title{A data set of gdp per state per year}
\format{
Data frame with columns
\describe{
\item{Year}{Year of observation}
\item{num_of_jobs}{The total number of people with a job in the state for the year}
\item{State}{The state of observation.}
}
}
\usage{
total_employment_data
}
\description{
A data set of gdp per state per year
}
\examples{
  total_employment_data
}
\keyword{datasets}
