% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unemployment.R
\name{unemployment}
\alias{unemployment}
\title{download Unemployment data}
\usage{
unemployment(
 
    url1 = "http://www.icip.iastate.edu/sites/default/files/uploads/tables/employment/emp-unemployment.xls"
)
}
\arguments{
\item{url1}{\itemize{
\item The URL pulls from Iowa state university data set of annual unemployment.
}}
}
\description{
download Unemployment data
}
