% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea.r
\name{agg_state_year}
\alias{agg_state_year}
\title{agg_state_year}
\usage{
agg_state_year(data)
}
\arguments{
\item{data}{\itemize{
\item the dataset from the bea.gov API that needs to be aggregated.
}}
}
\value{
returns the dataset aggregated to data values per state per year.
}
\description{
agg_state_year
}
