% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea2.r, R/hud.r
\name{ST_to_State}
\alias{ST_to_State}
\title{ST_to_State}
\usage{
ST_to_State(dataframe)

ST_to_State(dataframe)
}
\arguments{
\item{dataframe}{a df with a column named "state"}
}
\value{
The same dataframe but with the full statename

The same dataframe but with the full statename
}
\description{
ST_to_State

ST_to_State
}
